---------------------- version: 1.0.0+12
-- working on the grades page

----------------------
version: 1.0.0+11

-- fixed back position indicator and not order in exercise

-- attempting to fix order of arrival and undo that creates bugs in graph

-- fixed graphs in performance page

----------------------
version: 1.0.0+10

-- fixed many errors regrading web that came from moving to offline.

----------------------
version: 1.0.0+9

-- added matrix a new view to alonka exercise

-- updated comments dialog to save without adding and also added multiline

-- added swipe and arrows near the title in performance page

-- made work offline from start

-- added matrix to alonka graph

-- cleaned app (old images, files etc)

-- adding performance to the interview page,

-- added comments indicator to grades page

-- freezing the first col in the page grade

----------------------
version: 1.0.0+8

--- adding order of arrival grading to exercises

----------------------
version: 1.0.0+6

--  adding audio

----------------------
version: 1.0.0+5

--  Enabling Web

----------------------
version: 1.0.0+4

--- adding qualified recruits data when event finalized and aligning with admin

------------------------
version: 1.0.0+3

-- adding android id and manual device pairing.

--- name change and store config to support config.
------------------
dev 1,4
-----------
--added android id on log press of title in login

--adding kiosk support

-----
dev 1,3
-----
Tablet view added

-----
dev 1,2
-----

-- TODO  need to add version management admin for weighted avg

-- fixed the y-axis in Sakim chart so it is like Meshulash (inverted)

-- add grid only use case for Sakim and Meshulash

-- fixed save the kid AI report for future use

-- added #.## to grades

fixed - save button can not be reached by some phones.

fiexed - removed await in alonka round, so there is no interbnet issue

------------ done -----------------

-- sort fixes - list of instructors names dropdown in admin,
 list of events in Admin and Instructor

-- fixed some sizing and spacing

-- fixed alonka check of last round paricipants, when alonka starts it throws an error as last does not exist.

- added guide on docs and linked to app bar , for each page link to relevant screenshot

--------
dev 1.1
---------

-- TODO - save the report for future use

-- TODO  need to add version management admin for weighted avg

- added dropped to long press in comments page, and support for it in the relevant pages

- added grades for testes in perofrmace page
- fixing alonks GO to be viable only when prev round was finalized

- disabled editing of cols in grades page, only instructor grade is enabled

- fixed the Admin index, when del an event it del all events.

- added version to menu in home page and build script

- fixed p. ai report to work only on button click

- double click to decrease round for participant - sakim and meshulash

- added in Alonka an order by button, arrival or by order

- fixed col size in grade page

- finish weighted Avg task - move to cloud - moved

+++ lib/models/participant.dart
@@ -14,11 +14,11 @@ class Participant {
   double burGrade = 0;
   ParticipantStatus status = ParticipantStatus.Active;
   String fullName = '';
-  int instructorGrade = 0;
-  int instructorMeshulashGrade = 0; // Instructor's grade for Meshulash exercise
-  int instructorAlonkaGrade = 0; // Instructor's grade for Alonka exercise
-  int instructorSakimGrade = 0; // Instructor's grade for Sakim exercise
-  int instructorBurGrade = 0; // Instructor's grade for Bur exercise
+  double instructorGrade = 0.0;
+  double instructorMeshulashGrade = 0.0; // Instructor's grade for Meshulash exercise
+  double instructorAlonkaGrade = 0.0; // Instructor's grade for Alonka exercise
+  double instructorSakimGrade = 0.0; // Instructor's grade for Sakim exercise
+  int instructorBurGrade = 0; // Instructor's grade for Bur exercise (deprecated - use burGrades collection)
   double systemGrade = 0;
   String name = '';
   int groupNumber = 0;
@@ -34,7 +34,7 @@ class Participant {
   List<String> interviewInstructorComments = [];
 
   /// Set final instructor grade
-  setFinalGrade(int grade) {
+  setFinalGrade(double grade) {
     instructorGrade = grade;
   }
 
@@ -83,10 +83,10 @@ class Participant {
           (e) => e.toString().split('.').last == json['status'],
           orElse: () => ParticipantStatus.Active)
       ..fullName = json['fullName'] ?? ''
-      ..instructorGrade = json['instructorGrade'] ?? 0
-      ..instructorMeshulashGrade = json['instructorMeshulashGrade'] ?? 0
-      ..instructorAlonkaGrade = json['instructorAlonkaGrade'] ?? 0
-      ..instructorSakimGrade = json['instructorSakimGrade'] ?? 0
+      ..instructorGrade = (json['instructorGrade'] ?? 0).toDouble()
+      ..instructorMeshulashGrade = (json['instructorMeshulashGrade'] ?? 0).toDouble()
+      ..instructorAlonkaGrade = (json['instructorAlonkaGrade'] ?? 0).toDouble()
+      ..instructorSakimGrade = (json['instructorSakimGrade'] ?? 0).toDouble()
       ..instructorBurGrade = json['instructorBurGrade'] ?? 0
       ..systemGrade = (json['systemGrade'] ?? 0).toDouble()
       ..groupNumber = json['groupNumber'] ?? 0
